@@ -15,10 +15,10 @@
   "author": "demo",
   "license": "MIT",
   "dependencies": {
-    "jquery": "^3.1.1",
+    "jquery": "^1.11.3",
     "qwest": "^4.4.5",
-    "vue": "^1.0.26",
-    "vue-router": "^0.7.13"
+    "vue": "^2.1.10",
+    "vue-router": "^2.2.1"
   },
   "devDependencies": {
     "autoprefixer": "^6.5.1",
@@ -34,6 +34,7 @@
     "html-webpack-plugin": "^2.24.1",
     "image-webpack-loader": "^3.0.0",
     "ip": "^1.1.4",
+    "net": "^1.0.2",
     "node-sass": "^3.10.1",
     "open": "0.0.5",
     "postcss-loader": "^1.1.1",
@@ -41,10 +42,11 @@
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "vue-hot-reload-api": "^1.3.3",
-    "vue-html-loader": "^1.0.0",
-    "vue-loader": "^8.5.4",
-    "vue-style-loader": "^1.0.0",
+    "vue-hot-reload-api": "^2.0.8",
+    "vue-html-loader": "^1.2.3",
+    "vue-loader": "^11.0.0",
+    "vue-style-loader": "^2.0.0",
+    "vue-template-compiler": "^2.1.10",
     "webpack": "^1.13.3",
     "webpack-dev-server": "^1.16.2"
   }
