@@ -1,49 +1,47 @@
 <template>
+    <div>
+        <div class="demo1">
+            <h2><p><em>啦啦啦啦啦啦啦啦</em></p></h2>
 
-<div class="demo1">
-    <h2><p><em>啦啦啦啦啦啦啦啦</em></p></h2>
+            <h2><p><em>啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦</em></p></h2>
 
-    <h2><p><em>啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦</em></p></h2>
+            <h2><p><em>啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦</em></p></h2>
+        </div>
 
-    <h2><p><em>啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦啦</em></p></h2>
-</div>
+        <div class="demo2">
+            <ul>
+                <li>1</li>
+                <li>2</li>
+                <li>3</li>
+                <li>4</li>
+                <li>5</li>
+            </ul>
+        </div>
 
-<div class="demo2">
-    <ul>
-        <li>1</li>
-        <li>2</li>
-        <li>3</li>
-        <li>4</li>
-        <li>5</li>
-    </ul>
-</div>
-
-<div class="demo3">
-    <ul>
-        <li>1</li>
-        <li>2</li>
-        <li>3</li>
-        <li>4</li>
-        <li>5</li>
-    </ul>
-</div>
-
+        <div class="demo3">
+            <ul>
+                <li>1</li>
+                <li>2</li>
+                <li>3</li>
+                <li>4</li>
+                <li>5</li>
+            </ul>
+        </div>
+    </div>
 </template>
 
 <script>
-
 </script>
 
 <style scoped lang="scss">
+* {
+    margin: 0;
+    padding: 0;
+}
 
-    * {
-        margin: 0;
-        padding: 0;
-    }
-
-    div {
-        margin-bottom: 20px;
-    }
+div {
+    margin-bottom: 20px;
+}
 
 .demo1 {
     h2 {
@@ -52,13 +50,11 @@
         background: pink;
         margin-bottom: 20px;
     }
-
     p {
         display: inline-block;
         text-align: left;
         background: gray;
     }
-
     em {
         display: -webkit-box;
         // display: flex;
@@ -77,14 +73,12 @@
         height: 30px;
         line-height: 30px;
         overflow: hidden;
-
         &:after {
             content: '';
             width: 100%;
             display: inline-block;
         }
     }
-
     li {
         list-style: none;
         display: inline-block;
@@ -93,7 +87,6 @@
 
 .demo3 {
     overflow: hidden;
-
     ul {
         background: pink;
         height: 30px;
@@ -101,7 +94,6 @@
         overflow: hidden;
         margin-left: -1px;
     }
-
     li {
         box-sizing: border-box;
         list-style: none;
