@@ -1,30 +1,30 @@
 <template>
-
-    <div class="grid-test">
-        <div class="grid-row grid-gutter">
-            <div class="grid-col grid-col-8">
-                <p>col-8</p>
-            </div>
-            <div class="grid-col grid-col-8">
-                <p>col-8</p>
-            </div>
-            <div class="grid-col grid-col-8">
-                <p>col-8</p>
+    <div>
+        <div class="grid-test">
+            <div class="grid-row grid-gutter">
+                <div class="grid-col grid-col-8">
+                    <p>col-8</p>
+                </div>
+                <div class="grid-col grid-col-8">
+                    <p>col-8</p>
+                </div>
+                <div class="grid-col grid-col-8">
+                    <p>col-8</p>
+                </div>
             </div>
         </div>
+
+        <i class="iconfont"></i>
+        <i class="ion-xbox"></i>
+
+        <div class="test-bg"></div>
+
+        <ul class="counter">
+            <li>title</li>
+            <li>title</li>
+            <li>title</li>
+        </ul>
     </div>
-
-    <i class="iconfont"></i>
-    <i class="ion-xbox"></i>
-
-    <div class="test-bg"></div>
-
-    <ul class="counter">
-        <li>title</li>
-        <li>title</li>
-        <li>title</li>
-    </ul>
-
 </template>
 
 <script>
@@ -61,7 +61,6 @@ body {
 // minxin-grid
 $gridCol: 24;
 $gridGutter: 8px;
-
 @for $n from 1 through $gridCol {
     .grid-col-#{$n} {
         float: left;
@@ -82,6 +81,7 @@ $gridGutter: 8px;
         background: rgba(0, 123, 222, .5);
     }
 }
+
 .grid-col:nth-of-type(2n) {
     background: rgba(128, 128, 128, .3);
     p {
@@ -89,28 +89,24 @@ $gridGutter: 8px;
     }
 }
 
-
 // iconfont
 $iconfont-path: '../fonts/ionicons/';
 $iconfont-family: 'iconfont';
 $iconfont-version: 'v0.0.1';
-
 @font-face {
- font-family: $iconfont-family;
- src:url("#{$iconfont-path}/ionicons.eot?v=#{$iconfont-version}");
- src:url("#{$iconfont-path}/ionicons.eot?v=#{$iconfont-version}#iefix") format("embedded-opentype"),
-  url("#{$iconfont-path}/ionicons.ttf?v=#{$iconfont-version}") format("truetype"),
-  url("#{$iconfont-path}/ionicons.woff?v=#{$iconfont-version}") format("woff"),
-  url("#{$iconfont-path}/ionicons.svg?v=#{$iconfont-version}#Ionicons") format("svg");
+    font-family: $iconfont-family;
+    src:url("#{$iconfont-path}/ionicons.eot?v=#{$iconfont-version}");
+    src:url("#{$iconfont-path}/ionicons.eot?v=#{$iconfont-version}#iefix") format("embedded-opentype"),
+    url("#{$iconfont-path}/ionicons.ttf?v=#{$iconfont-version}") format("truetype"),
+    url("#{$iconfont-path}/ionicons.woff?v=#{$iconfont-version}") format("woff"),
+    url("#{$iconfont-path}/ionicons.svg?v=#{$iconfont-version}#Ionicons") format("svg");
 }
-
 
 .iconfont {
     font-family: $iconfont-family;
     font-weight: 400;
     font-style: normal;
     display: inline-block;
-
     &:before {
         content: '\f25c';
     }
@@ -124,11 +120,9 @@ $iconfont-version: 'v0.0.1';
 
 .counter {
     counter-reset: test test2;
-
     li:before {
         counter-increment: test test2;
         content: counter(test) '.' counter(test2);
     }
 }
-
 </style>
