@@ -1,130 +1,129 @@
 <template>
+    <div>
+        <ul>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+            <li class="svg"></li>
+        </ul>
 
-    <ul>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-        <li class="svg"></li>
-    </ul>
+        <hr>
 
-    <hr>
-
-    <svg width="32" height="1040" viewBox="0 0 32 1040">
-    <svg id="icon-google-plus2" width="32" height="32" viewBox="0 0 32 32" x="0" y="0">
-    <path fill="#ef443d" d="M0.025 27.177c-0.008-0.079-0.014-0.158-0.018-0.238 0.004 0.080 0.011 0.159 0.018 0.238zM7.372 17.661c2.875 0.086 4.804-2.897 4.308-6.662s-3.231-6.787-6.106-6.873c-2.876-0.085-4.804 2.796-4.308 6.562 0.496 3.765 3.23 6.887 6.106 6.973zM32 8v-2.666c0-2.934-2.399-5.334-5.333-5.334h-21.333c-2.884 0-5.25 2.32-5.33 5.185 1.824-1.606 4.354-2.947 6.965-2.947 2.791 0 11.164 0 11.164 0l-2.498 2.113h-3.54c2.348 0.9 3.599 3.629 3.599 6.429 0 2.351-1.307 4.374-3.153 5.812-1.801 1.403-2.143 1.991-2.143 3.184 0 1.018 1.93 2.75 2.938 3.462 2.949 2.079 3.904 4.010 3.904 7.233 0 0.513-0.064 1.026-0.19 1.53h9.617c2.934 0 5.333-2.398 5.333-5.334v-16.666h-6v6h-2v-6h-6v-2h6v-6h2v6h6zM5.809 23.936c0.675 0 1.294-0.018 1.936-0.018-0.848-0.823-1.52-1.831-1.52-3.074 0-0.738 0.236-1.448 0.567-2.079-0.337 0.024-0.681 0.031-1.035 0.031-2.324 0-4.297-0.752-5.756-1.995v2.101l0 6.304c1.67-0.793 3.653-1.269 5.809-1.269zM0.107 27.727c-0.035-0.171-0.061-0.344-0.079-0.52 0.018 0.176 0.045 0.349 0.079 0.52zM14.233 29.776c-0.471-1.838-2.139-2.749-4.465-4.361-0.846-0.273-1.778-0.434-2.778-0.444-2.801-0.030-5.41 1.092-6.882 2.762 0.498 2.428 2.657 4.267 5.226 4.267h8.951c0.057-0.348 0.084-0.707 0.084-1.076 0-0.392-0.048-0.775-0.137-1.148z"/>
-    </svg>
-    <svg id="icon-facebook2" width="32" height="32" viewBox="0 0 32 32" x="0" y="42">
-    <path fill="#3c5ea8" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.387 5.334 5.333 5.334h10.667v-14h-4v-4h4v-3c0-2.761 2.239-5 5-5h5v4h-5c-0.552 0-1 0.448-1 1v3h5.5l-1 4h-4.5v14h6.667c2.945 0 5.333-2.388 5.333-5.334v-21.332c0-2.946-2.387-5.334-5.333-5.334z"/>
-    </svg>
-    <svg id="icon-twitter2" width="32" height="32" viewBox="0 0 32 32" x="0" y="84">
-    <path fill="#02bceb" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM23.952 11.921c0.008 0.176 0.012 0.353 0.012 0.531 0 5.422-4.127 11.675-11.675 11.675-2.317 0-4.474-0.679-6.29-1.844 0.321 0.038 0.648 0.058 0.979 0.058 1.922 0 3.692-0.656 5.096-1.757-1.796-0.033-3.311-1.219-3.833-2.849 0.251 0.048 0.508 0.074 0.772 0.074 0.374 0 0.737-0.050 1.081-0.144-1.877-0.377-3.291-2.035-3.291-4.023 0-0.017 0-0.034 0-0.052 0.553 0.307 1.186 0.492 1.858 0.513-1.101-0.736-1.825-1.992-1.825-3.415 0-0.752 0.202-1.457 0.556-2.063 2.024 2.482 5.047 4.116 8.457 4.287-0.070-0.3-0.106-0.614-0.106-0.935 0-2.266 1.837-4.103 4.103-4.103 1.18 0 2.247 0.498 2.995 1.296 0.935-0.184 1.813-0.525 2.606-0.996-0.306 0.958-0.957 1.762-1.804 2.27 0.83-0.099 1.621-0.32 2.357-0.646-0.55 0.823-1.245 1.545-2.047 2.124z"/>
-    </svg>
-    <svg id="icon-feed3" width="32" height="32" viewBox="0 0 32 32" x="0" y="126">
-    <path fill="#f37e0d" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.935 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM8.717 25.977c-1.499 0-2.717-1.209-2.717-2.707 0-1.49 1.218-2.713 2.717-2.713 1.505 0 2.721 1.223 2.721 2.713-0 1.498-1.216 2.707-2.721 2.707zM15.546 26c0-2.557-0.994-4.963-2.796-6.764-1.804-1.805-4.199-2.801-6.747-2.801v-3.912c7.427 0 13.473 6.045 13.473 13.477l-3.93 0zM22.488 26c0-9.096-7.396-16.498-16.48-16.498v-3.914c11.25 0 20.405 9.16 20.405 20.412l-3.924-0z"/>
-    </svg>
-    <svg id="icon-youtube3" width="32" height="32" viewBox="0 0 32 32" x="0" y="168">
-    <path fill="#de4943" d="M9.162 0l-2.162 4.289-2.162-4.289h-2.758l3.897 6.791 0.023-0.013v5.222h2v-5.222l0.023 0.013 3.897-6.791z"/>
-    <path fill="#de4943" d="M15 4c0.542 0 1 0.458 1 1v4c0 0.542-0.458 1-1 1s-1-0.458-1-1v-4c0-0.542 0.458-1 1-1zM15 2c-1.65 0-3 1.35-3 3v4c0 1.65 1.35 3 3 3s3-1.35 3-3v-4c0-1.65-1.35-3-3-3v0z"/>
-    <path fill="#de4943" d="M24 2v7.874c-0.873 0.728-2 1.167-2-1.426v-6.448h-2v6.854h0.003c0.031 1.653 0.379 4.63 3.997 2.21v0.936h2v-10h-2z"/>
-    <path fill="#de4943" d="M27 22c-0.551 0-1 0.449-1 1v1h2v-1c0-0.551-0.449-1-1-1z"/>
-    <path fill="#de4943" d="M18 23v5.25c0.677 0.677 2 0.75 2-0.25s0-3.688 0-4.688-1-1.313-2-0.313z"/>
-    <path fill="#de4943" d="M31.836 18.173c-0.087-1.152-0.501-2.092-1.243-2.82s-1.69-1.125-2.846-1.19c-2.181-0.109-7.17-0.163-11.554-0.163s-9.728 0.054-11.909 0.163c-1.156 0.065-2.104 0.462-2.846 1.19s-1.156 1.668-1.243 2.82c-0.131 2.347-0.196 3.198-0.196 4.372s0.065 2.934 0.196 5.282c0.087 1.152 0.502 2.092 1.243 2.82s1.69 1.125 2.846 1.19c2.181 0.109 7.525 0.163 11.909 0.163s9.373-0.054 11.554-0.163c1.156-0.065 2.104-0.462 2.846-1.19s1.156-1.668 1.243-2.82c0.109-1.956 0.164-3.717 0.164-5.282s-0.055-2.416-0.164-4.372zM6 30h-2v-10h-2v-2h6v2h-2v10zM14 30h-2v-0.936c-3.802 2.134-3.966 0.069-3.997-1.585h-0.003v-5.479h2v5.51c0 1.208 1.127 1.091 2 0.364v-5.874h2v8zM22 27.971c0 2.090-2.141 2.7-4 1.075v0.954h-2v-12h2v3.25c2-2 4-1.25 4 0.75s0 3.787 0 5.971zM30 24v1h-4v2c0 0.551 0.449 1 1 1s1-0.449 1-1v-1h2v1c0 1.654-1.346 3-3 3s-3-1.346-3-3v-4c0-1.654 1.346-3 3-3s3 1.346 3 3v1z"/>
-    </svg>
-    <svg id="icon-vimeo2" width="32" height="32" viewBox="0 0 32 32" x="0" y="210">
-    <path fill="#04ae7f" d="M26.666 0h-21.332c-2.934 0-5.334 2.401-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.332c2.934 0 5.334-2.398 5.334-5.334v-21.332c0-2.934-2.4-5.334-5.334-5.334zM25.402 13.691c-1.061 5.978-6.991 11.041-8.774 12.197s-3.411-0.463-4.001-1.687c-0.675-1.397-2.7-8.967-3.23-9.594-0.53-0.626-2.121 0.627-2.121 0.627l-0.772-1.012c0 0 3.231-3.858 5.689-4.34 2.606-0.511 2.603 4.002 3.229 6.508 0.606 2.424 1.014 3.811 1.543 3.811 0.53 0 1.543-1.352 2.652-3.424 1.111-2.074-0.048-3.907-2.218-2.603 0.868-5.208 9.064-6.46 8.004-0.483z"/>
-    </svg>
-    <svg id="icon-lanyrd" width="32" height="32" viewBox="0 0 32 32" x="0" y="252">
-    <path fill="#ef6db3" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM26.365 25.24c-3.542 1.167-10.458 3.438-11.75 3.76s-2.625 0.865-3.375-1.302c0 0-6.375-18.833-6.958-20.5s-1.151-2.671 0.167-3.198c3.043-1.215 6.208-2.766 6.792-1 0.522 1.58 6.083 19 6.083 19s6.542-2.302 8.333-2.802 2.406-0.833 2.875 1.625 1.375 3.25-2.167 4.417z"/>
-    </svg>
-    <svg id="icon-flickr3" width="32" height="32" viewBox="0 0 32 32" x="0" y="294">
-    <path fill="#e31884" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-21.332c0-2.933-2.399-5.334-5.332-5.334zM9 21c-2.761 0-5-2.239-5-5s2.239-5 5-5 5 2.239 5 5c0 2.761-2.239 5-5 5zM23 21c-2.761 0-5-2.239-5-5s2.239-5 5-5 5 2.239 5 5c0 2.761-2.239 5-5 5z"/>
-    </svg>
-    <svg id="icon-picassa2" width="32" height="32" viewBox="0 0 32 32" x="0" y="336">
-    <path fill="#fa8526" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-21.332c0-2.933-2.399-5.334-5.332-5.334zM16 2c2.147 0 4.181 0.484 6 1.348v8.973l-9.799-9.799c1.208-0.34 2.482-0.522 3.799-0.522zM2 16c0-5.497 3.168-10.252 7.777-12.543l3.543 3.543-11.155 11.155c-0.109-0.703-0.165-1.422-0.165-2.155zM8 27.489c-2.334-1.628-4.149-3.948-5.146-6.664l5.146-5.146v11.81zM16 30c-2.147 0-4.181-0.484-6-1.348v-4.652h17.489c-2.53 3.627-6.732 6-11.489 6zM28.652 22h-4.652v-17.489c3.627 2.53 6 6.732 6 11.489 0 2.147-0.484 4.181-1.348 6z"/>
-    </svg>
-    <svg id="icon-dribbble2" width="32" height="32" viewBox="0 0 32 32" x="0" y="378">
-    <path fill="#ef6797" d="M16.836 16.123c-6.167 1.994-9.449 7.427-9.72 7.891-1.91-2.12-3.074-4.924-3.074-7.996 0-0.122 0.002-0.244 0.006-0.366 0.523 0.011 6.318 0.085 12.29-1.702 0.342 0.67 0.669 1.35 0.97 2.029-0.158 0.045-0.315 0.092-0.471 0.142zM23.927 7.073c-0.215 0.306-2.015 2.726-6.267 4.32-1.96-3.6-4.131-6.558-4.462-7 0.899-0.217 1.837-0.332 2.802-0.332 3.038 0 5.815 1.14 7.927 3.013zM15.35 12.112c-5.594 1.485-10.52 1.464-11.055 1.457 0.775-3.711 3.276-6.795 6.622-8.373 0.315 0.432 2.452 3.394 4.434 6.916zM18.041 17.984c0.026-0.009 0.052-0.018 0.079-0.027 1.679 4.363 2.373 8.019 2.551 9.068-1.436 0.611-3.015 0.95-4.671 0.95-2.764 0-5.313-0.943-7.34-2.524 0.215-0.44 2.636-5.117 9.381-7.468zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM16 30c-7.72 0-14-6.28-14-14s6.28-14 14-14c7.72 0 14 6.28 14 14s-6.28 14-14 14zM20.358 17.405c3.747-0.6 7.040 0.382 7.449 0.511-0.533 3.327-2.443 6.201-5.126 8.015-0.123-0.727-0.759-4.228-2.323-8.526zM18.579 13.171c4.421-1.805 6.425-4.374 6.677-4.715 1.665 2.035 2.674 4.627 2.7 7.449-0.394-0.083-4.348-0.885-8.327-0.385-0.084-0.205-0.17-0.41-0.258-0.617-0.248-0.583-0.515-1.162-0.793-1.732z"/>
-    </svg>
-    <svg id="icon-forrst2" width="32" height="32" viewBox="0 0 32 32" x="0" y="420">
-    <path fill="#d61e1e" d="M14 18h4v8h-4zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM18 26v4h-4v-4h-8l10-22 10 22h-8z"/>
-    </svg>
-    <svg id="icon-deviantart2" width="32" height="32" viewBox="0 0 32 32" x="0" y="462">
-    <path fill="#0b6e5c" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM30 24h-12v-7.375c0 0 2.198 0.031 3.375 1.25v2.125h4.825c-0.369-1.184-1.158-2.347-2.285-3.333-1.937-1.695-5.1-2.542-7.915-2.542v9.875h-14v-3c0-1.356 0.458-2.704 1.324-3.899 0.769-1.060 1.84-1.995 3.183-2.779 1.576-0.919 3.455-1.585 5.493-1.962v-4.361h4v4c4.594 0 6.925 0.825 9.493 2.323 1.344 0.784 2.415 1.719 3.183 2.779 0.866 1.194 1.324 2.543 1.324 3.899v3zM6.237 20c0.369-1.184 1.158-2.347 2.285-3.333 0.986-0.862 2.176-1.537 3.478-1.992v5.326h-5.763z"/>
-    </svg>
-    <svg id="icon-steam2" width="32" height="32" viewBox="0 0 32 32" x="0" y="504">
-    <path fill="#c45d3e" d="M9.498 26.125c0.848 0 1.681-0.41 2.187-1.169 0.804-1.206 0.478-2.836-0.728-3.64l-2.065-1.377c0.353-0.096 0.725-0.149 1.109-0.149 2.325 0 4.209 1.884 4.209 4.209s-1.885 4.209-4.209 4.209c-2.263 0-4.108-1.785-4.205-4.024l2.249 1.499c0.448 0.298 0.953 0.441 1.454 0.441zM26.667 0c2.934 0 5.333 2.4 5.333 5.334v21.332c0 2.936-2.399 5.334-5.333 5.334h-21.333c-2.934 0-5.334-2.398-5.334-5.334v-6.345l3.792 2.528c-0.364 1.974 0.217 4.088 1.744 5.615 2.466 2.466 6.463 2.466 8.929 0 1.515-1.515 2.098-3.607 1.752-5.568l7.783-6.968c1.445-0.207 2.838-0.867 3.95-1.979 2.734-2.734 2.734-7.166 0-9.899s-7.166-2.734-9.9 0c-1.112 1.112-1.771 2.505-1.979 3.95v0l-6.397 9.695c-0.995 0.051-1.981 0.336-2.866 0.856l-6.808-4.539v-8.678c0-2.934 2.4-5.334 5.334-5.334h21.333zM28 9c0-2.761-2.239-5-5-5s-5 2.239-5 5 2.239 5 5 5 5-2.239 5-5zM20 9c0-1.657 1.343-3 3-3s3 1.343 3 3-1.343 3-3 3-3-1.343-3-3z"/>
-    </svg>
-    <svg id="icon-github" width="32" height="32" viewBox="0 0 32 32" x="0" y="546">
-    <path fill="#484848" d="M16 0.395c-8.836 0-16 7.163-16 16 0 7.069 4.585 13.067 10.942 15.182 0.8 0.148 1.094-0.347 1.094-0.77 0-0.381-0.015-1.642-0.022-2.979-4.452 0.968-5.391-1.888-5.391-1.888-0.728-1.849-1.776-2.341-1.776-2.341-1.452-0.993 0.11-0.973 0.11-0.973 1.606 0.113 2.452 1.649 2.452 1.649 1.427 2.446 3.743 1.739 4.656 1.33 0.143-1.034 0.558-1.74 1.016-2.14-3.554-0.404-7.29-1.777-7.29-7.907 0-1.747 0.625-3.174 1.649-4.295-0.166-0.403-0.714-2.030 0.155-4.234 0 0 1.344-0.43 4.401 1.64 1.276-0.355 2.645-0.532 4.005-0.539 1.359 0.006 2.729 0.184 4.008 0.539 3.054-2.070 4.395-1.64 4.395-1.64 0.871 2.204 0.323 3.831 0.157 4.234 1.026 1.12 1.647 2.548 1.647 4.295 0 6.145-3.743 7.498-7.306 7.895 0.574 0.497 1.085 1.47 1.085 2.963 0 2.141-0.019 3.864-0.019 4.391 0 0.426 0.288 0.925 1.099 0.768 6.354-2.118 10.933-8.113 10.933-15.18 0-8.837-7.164-16-16-16z"/>
-    </svg>
-    <svg id="icon-wordpress2" width="32" height="32" viewBox="0 0 32 32" x="0" y="588">
-    <path fill="#4a4a4a" d="M16.211 17.022l3.688 9.842c0.024 0.058 0.054 0.111 0.086 0.161-1.247 0.428-2.588 0.663-3.985 0.663-1.178 0-2.315-0.168-3.39-0.476l3.601-10.189zM28 16c0 4.312-2.399 8.076-5.967 10.102l3.665-10.321c0.685-1.667 0.913-3.001 0.913-4.186 0-0.43-0.029-0.83-0.081-1.202 0.937 1.664 1.47 3.575 1.47 5.607zM4 16c0-1.694 0.373-3.303 1.039-4.756l5.724 15.274c-4.003-1.894-6.763-5.892-6.763-10.518zM5.974 9.58c2.146-3.173 5.834-5.268 10.026-5.268 3.124 0 5.968 1.163 8.103 3.069-0.052-0.003-0.102-0.009-0.155-0.009-1.179 0-2.015 1-2.015 2.074 0 0.963 0.571 1.778 1.179 2.741 0.456 0.778 0.989 1.778 0.989 3.223 0 1-0.395 2.161-0.913 3.778l-1.197 3.895-4.337-12.563c0.722-0.037 1.373-0.111 1.373-0.111 0.647-0.075 0.571-1-0.076-0.963 0 0-1.944 0.149-3.198 0.149-1.179 0-3.16-0.149-3.16-0.149-0.647-0.037-0.723 0.926-0.076 0.963 0 0 0.612 0.074 1.259 0.111l1.869 4.989-2.626 7.671-4.37-12.66c0.723-0.037 1.373-0.111 1.373-0.111 0.646-0.075 0.57-1-0.077-0.963 0 0-1.943 0.149-3.198 0.149-0.225 0-0.491-0.005-0.772-0.014zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM30 16c0 7.732-6.268 14-14 14s-14-6.268-14-14 6.268-14 14-14 14 6.268 14 14z"/>
-    </svg>
-    <svg id="icon-blogger2" width="32" height="32" viewBox="0 0 32 32" x="0" y="630">
-    <path fill="#f17c1d" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.388 5.334 5.333 5.334h21.334c2.945 0 5.333-2.388 5.333-5.334l0-21.332c0-2.946-2.388-5.334-5.333-5.334zM28 20.253c0 4.281-3.484 7.747-7.795 7.747h-8.401c-4.308 0-7.804-3.466-7.804-7.747v-8.503c0.003-4.281 3.496-7.75 7.804-7.75h3.933c4.311 0 7.762 3.219 7.762 7.5 0.056 0.803 0.779 1.5 1.603 1.5h1.345c0.862 0 1.552 0.904 1.552 1.76v5.493zM22 20c0 1.1-0.9 2-2 2h-8c-1.1 0-2-0.9-2-2v0c0-1.1 0.9-2 2-2h8c1.1 0 2 0.9 2 2v0zM18 12c0 1.1-0.9 2-2 2h-4c-1.1 0-2-0.9-2-2v0c0-1.1 0.9-2 2-2h4c1.1 0 2 0.9 2 2v0z"/>
-    </svg>
-    <svg id="icon-tumblr2" width="32" height="32" viewBox="0 0 32 32" x="0" y="672">
-    <path fill="#3c576d" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.933 0 5.332-2.398 5.332-5.334v-21.332c-0-2.933-2.399-5.334-5.332-5.334zM22.866 25.771c-0.942 0.443-1.798 0.756-2.563 0.936-0.765 0.178-1.593 0.267-2.481 0.267-1.010 0-1.605-0.127-2.381-0.381-0.775-0.256-1.438-0.621-1.984-1.090-0.549-0.473-0.928-0.975-1.14-1.506s-0.317-1.303-0.317-2.313v-7.744h-3v-3.127c0.867-0.281 1.873-0.685 2.49-1.211 0.62-0.527 1.116-1.158 1.49-1.896 0.375-0.736 0.633-1.676 0.774-2.815h3.141v5.108h5.105v3.941h-5.106v5.662c0 1.281-0.017 2.020 0.119 2.383 0.135 0.361 0.473 0.736 0.841 0.953 0.489 0.293 1.047 0.439 1.676 0.439 1.118 0 2.231-0.363 3.336-1.090v3.482z"/>
-    </svg>
-    <svg id="icon-soundcloud2" width="32" height="32" viewBox="0 0 32 32" x="0" y="714">
-    <path fill="#dbbaa8" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v14.666l0.5-2h1l0.5 2-0.5 2h-1l-0.5-2v6.666c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-6.631c0 2.19-1.846 3.965-4.125 3.965-0.074 0-11.424-0.006-11.434-0.007-0.246-0.024-0.441-0.226-0.441-0.469v-12.585c0-0.231 0.082-0.35 0.393-0.466 0.801-0.298 1.704-0.474 2.635-0.474 3.803 0 6.922 2.803 7.25 6.377 0.491-0.198 1.031-0.308 1.596-0.308 2.278 0 4.126 1.776 4.126 3.966v-14.701c0-2.934-2.399-5.334-5.332-5.334zM5.5 24h-1l-0.5-4 0.5-4h1l0.5 4-0.5 4zM9.5 24h-1l-0.5-5.086 0.5-4.914h1l0.5 5-0.5 5zM13.5 24h-1l-0.5-7.007 0.5-6.993h1l0.5 6.993-0.5 7.007z"/>
-    </svg>
-    <svg id="icon-linkedin" width="32" height="32" viewBox="0 0 32 32" x="0" y="756">
-    <path fill="#3c91b3" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM12 26h-4v-14h4v14zM10 10c-1.105 0-2-0.895-2-2s0.895-2 2-2 2 0.895 2 2-0.895 2-2 2zM26 26h-4v-8c0-1.105-0.895-2-2-2s-2 0.895-2 2v8h-4v-14h4v2.483c0.825-1.133 2.086-2.483 3.5-2.483 2.485 0 4.5 2.239 4.5 5v9z"/>
-    </svg>
-    <svg id="icon-lastfm2" width="32" height="32" viewBox="0 0 32 32" x="0" y="798">
-    <path fill="#b8bb8c" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM23.329 23.778c-5.549 0-7.473-2.502-8.499-5.613l-1.027-3.207c-0.77-2.341-1.668-4.169-4.49-4.169-1.957 0-3.945 1.411-3.945 5.356 0 3.079 1.572 5.003 3.785 5.003 2.501 0 4.169-1.86 4.169-1.86l1.027 2.79c0 0-1.732 1.7-5.356 1.7-4.49 0-6.992-2.63-6.992-7.505 0-5.068 2.502-8.050 7.216-8.050 4.266 0 6.415 1.54 7.762 5.709l1.058 3.207c0.77 2.342 2.117 4.041 5.356 4.041 2.181 0 3.335-0.481 3.335-1.668 0-0.93-0.545-1.604-2.181-1.989l-2.181-0.513c-2.662-0.641-3.72-2.021-3.72-4.202 0-3.496 2.823-4.587 5.709-4.587 3.272 0 5.26 1.187 5.517 4.073l-3.207 0.385c-0.128-1.379-0.962-1.957-2.502-1.957-1.411 0-2.277 0.642-2.277 1.732 0 0.962 0.417 1.54 1.828 1.86l2.053 0.449c2.758 0.642 4.234 1.989 4.234 4.587-0 3.207-2.694 4.426-6.672 4.426z"/>
-    </svg>
-    <svg id="icon-delicious" width="32" height="32" viewBox="0 0 32 32" x="0" y="840">
-    <path fill="#b1110f" d="M26.668 0h-21.334c-2.934 0-5.334 2.401-5.334 5.334v21.332c0 2.935 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.398 5.332-5.334v-21.332c0-2.934-2.398-5.334-5.332-5.334zM30 26.666c0 0.885-0.348 1.721-0.98 2.354s-1.467 0.98-2.352 0.98h-10.668v-14h-14v-10.666c0-0.885 0.348-1.72 0.981-2.353s1.468-0.981 2.353-0.981h10.666v14h14v10.666z"/>
-    </svg>
-    <svg id="icon-stumbleupon" width="32" height="32" viewBox="0 0 32 32" x="0" y="882">
-    <path fill="#60bd0d" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM16 10c-1.103 0-2 0.897-2 2v8c0 3.308-2.692 6-6 6s-6-2.692-6-6v-4h4v4c0 1.103 0.897 2 2 2s2-0.897 2-2v-8c0-3.308 2.692-6 6-6s6 2.692 6 5.563v1.938l-2.563 0.75-1.438-0.75v-1.938c0-0.665-0.897-1.563-2-1.563zM30 20c0 3.308-2.692 6-6 6s-6-2.692-6-6.438v-3.875l1.438 0.75 2.563-0.75v3.875c0 1.54 0.897 2.438 2 2.438s2-0.897 2-2v-4h4v4z"/>
-    </svg>
-    <svg id="icon-pinterest2" width="32" height="32" viewBox="0 0 32 32" x="0" y="924">
-    <path fill="#2f6e95" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.387 5.334 5.333 5.334h21.334c2.945 0 5.333-2.388 5.333-5.334v-21.332c0-2.946-2.387-5.334-5.333-5.334zM17.915 25.126c-1.621 0-3.145-0.842-3.667-1.837 0 0-0.802 3.055-0.997 3.803-0.361 1.39-1.336 3.132-1.989 4.195l-1.093-0.387c-0.14-1.266-0.266-3.208 0.055-4.59 0.291-1.249 1.876-7.953 1.876-7.953s-0.479-0.958-0.479-2.375c0-2.225 1.29-3.886 2.895-3.886 1.365 0 2.025 1.025 2.025 2.254 0 1.373-0.874 3.425-1.325 5.327-0.377 1.593 0.799 2.892 2.369 2.892 2.844 0 5.030-2.999 5.030-7.327 0-3.831-2.753-6.509-6.683-6.509-4.552 0-7.225 3.415-7.225 6.943 0 1.375 0.53 2.85 1.191 3.651 0.131 0.158 0.15 0.297 0.111 0.459-0.121 0.506-0.391 1.593-0.444 1.815-0.070 0.293-0.232 0.355-0.535 0.214-1.998-0.93-3.248-3.852-3.248-6.198 0-5.047 3.667-9.682 10.572-9.682 5.55 0 9.864 3.955 9.864 9.241 0 5.514-3.477 9.952-8.302 9.952z"/>
-    </svg>
-    <svg id="icon-xing" width="32" height="32" viewBox="0 0 32 32" x="0" y="966">
-    <path fill="#914e93" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM9.743 20.837h-3.477c-0.209 0-0.366-0.094-0.454-0.237-0.092-0.149-0.096-0.341 0-0.533l3.694-6.52c0.004-0.007 0.004-0.012 0-0.020l-2.351-4.070c-0.097-0.194-0.111-0.385-0.020-0.533 0.089-0.144 0.265-0.217 0.474-0.217h3.477c0.533 0 0.795 0.344 0.968 0.652 0 0 2.376 4.146 2.391 4.168-0.141 0.248-3.754 6.638-3.754 6.638-0.178 0.321-0.428 0.672-0.948 0.672zM26.185 4.704l-7.698 13.612c-0.005 0.007-0.005 0.016 0 0.023l4.902 8.956c0.097 0.193 0.1 0.389 0.007 0.537-0.088 0.143-0.252 0.222-0.461 0.222h-3.473c-0.533 0-0.799-0.354-0.971-0.662 0 0-4.928-9.040-4.941-9.064 0.246-0.436 7.737-13.72 7.737-13.72 0.186-0.335 0.412-0.662 0.931-0.662h3.513c0.209 0 0.374 0.079 0.462 0.222 0.092 0.149 0.089 0.344-0.007 0.537z"/>
-    </svg>
-    <svg id="icon-flattr" width="32" height="32" viewBox="0 0 32 32" x="0" y="1008">
-    <path fill="#b9c5bf" d="M11.486 0c-7.605 0-11.486 4.38-11.486 12.558v0 17.157l7.45-7.457v-8.716c0-3.388 0.898-5.544 3.91-6.028v0c1.052-0.206 3.242-0.134 4.635-0.134v0 5.175c0 0.047 0.006 0.131 0.019 0.175v0c0.058 0.21 0.248 0.363 0.472 0.363v0c0.127 0 0.246-0.066 0.368-0.187v0l12.91-12.905-18.277-0.002zM24.55 9.742v8.715c0 3.388-0.898 5.544-3.91 6.028v0c-1.052 0.206-3.242 0.134-4.635 0.134v0-5.175c0-0.047-0.006-0.131-0.019-0.174v0c-0.058-0.21-0.248-0.363-0.472-0.363v0c-0.127-0-0.246 0.066-0.368 0.186v0l-12.91 12.905 18.277 0.002c7.605 0 11.486-4.38 11.486-12.558v0-17.157l-7.45 7.457z"/>
-    </svg>
-    </svg>
-
+        <svg width="32" height="1040" viewBox="0 0 32 1040">
+            <svg id="icon-google-plus2" width="32" height="32" viewBox="0 0 32 32" x="0" y="0">
+                <path fill="#ef443d" d="M0.025 27.177c-0.008-0.079-0.014-0.158-0.018-0.238 0.004 0.080 0.011 0.159 0.018 0.238zM7.372 17.661c2.875 0.086 4.804-2.897 4.308-6.662s-3.231-6.787-6.106-6.873c-2.876-0.085-4.804 2.796-4.308 6.562 0.496 3.765 3.23 6.887 6.106 6.973zM32 8v-2.666c0-2.934-2.399-5.334-5.333-5.334h-21.333c-2.884 0-5.25 2.32-5.33 5.185 1.824-1.606 4.354-2.947 6.965-2.947 2.791 0 11.164 0 11.164 0l-2.498 2.113h-3.54c2.348 0.9 3.599 3.629 3.599 6.429 0 2.351-1.307 4.374-3.153 5.812-1.801 1.403-2.143 1.991-2.143 3.184 0 1.018 1.93 2.75 2.938 3.462 2.949 2.079 3.904 4.010 3.904 7.233 0 0.513-0.064 1.026-0.19 1.53h9.617c2.934 0 5.333-2.398 5.333-5.334v-16.666h-6v6h-2v-6h-6v-2h6v-6h2v6h6zM5.809 23.936c0.675 0 1.294-0.018 1.936-0.018-0.848-0.823-1.52-1.831-1.52-3.074 0-0.738 0.236-1.448 0.567-2.079-0.337 0.024-0.681 0.031-1.035 0.031-2.324 0-4.297-0.752-5.756-1.995v2.101l0 6.304c1.67-0.793 3.653-1.269 5.809-1.269zM0.107 27.727c-0.035-0.171-0.061-0.344-0.079-0.52 0.018 0.176 0.045 0.349 0.079 0.52zM14.233 29.776c-0.471-1.838-2.139-2.749-4.465-4.361-0.846-0.273-1.778-0.434-2.778-0.444-2.801-0.030-5.41 1.092-6.882 2.762 0.498 2.428 2.657 4.267 5.226 4.267h8.951c0.057-0.348 0.084-0.707 0.084-1.076 0-0.392-0.048-0.775-0.137-1.148z" />
+            </svg>
+            <svg id="icon-facebook2" width="32" height="32" viewBox="0 0 32 32" x="0" y="42">
+                <path fill="#3c5ea8" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.387 5.334 5.333 5.334h10.667v-14h-4v-4h4v-3c0-2.761 2.239-5 5-5h5v4h-5c-0.552 0-1 0.448-1 1v3h5.5l-1 4h-4.5v14h6.667c2.945 0 5.333-2.388 5.333-5.334v-21.332c0-2.946-2.387-5.334-5.333-5.334z" />
+            </svg>
+            <svg id="icon-twitter2" width="32" height="32" viewBox="0 0 32 32" x="0" y="84">
+                <path fill="#02bceb" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM23.952 11.921c0.008 0.176 0.012 0.353 0.012 0.531 0 5.422-4.127 11.675-11.675 11.675-2.317 0-4.474-0.679-6.29-1.844 0.321 0.038 0.648 0.058 0.979 0.058 1.922 0 3.692-0.656 5.096-1.757-1.796-0.033-3.311-1.219-3.833-2.849 0.251 0.048 0.508 0.074 0.772 0.074 0.374 0 0.737-0.050 1.081-0.144-1.877-0.377-3.291-2.035-3.291-4.023 0-0.017 0-0.034 0-0.052 0.553 0.307 1.186 0.492 1.858 0.513-1.101-0.736-1.825-1.992-1.825-3.415 0-0.752 0.202-1.457 0.556-2.063 2.024 2.482 5.047 4.116 8.457 4.287-0.070-0.3-0.106-0.614-0.106-0.935 0-2.266 1.837-4.103 4.103-4.103 1.18 0 2.247 0.498 2.995 1.296 0.935-0.184 1.813-0.525 2.606-0.996-0.306 0.958-0.957 1.762-1.804 2.27 0.83-0.099 1.621-0.32 2.357-0.646-0.55 0.823-1.245 1.545-2.047 2.124z" />
+            </svg>
+            <svg id="icon-feed3" width="32" height="32" viewBox="0 0 32 32" x="0" y="126">
+                <path fill="#f37e0d" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.935 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM8.717 25.977c-1.499 0-2.717-1.209-2.717-2.707 0-1.49 1.218-2.713 2.717-2.713 1.505 0 2.721 1.223 2.721 2.713-0 1.498-1.216 2.707-2.721 2.707zM15.546 26c0-2.557-0.994-4.963-2.796-6.764-1.804-1.805-4.199-2.801-6.747-2.801v-3.912c7.427 0 13.473 6.045 13.473 13.477l-3.93 0zM22.488 26c0-9.096-7.396-16.498-16.48-16.498v-3.914c11.25 0 20.405 9.16 20.405 20.412l-3.924-0z" />
+            </svg>
+            <svg id="icon-youtube3" width="32" height="32" viewBox="0 0 32 32" x="0" y="168">
+                <path fill="#de4943" d="M9.162 0l-2.162 4.289-2.162-4.289h-2.758l3.897 6.791 0.023-0.013v5.222h2v-5.222l0.023 0.013 3.897-6.791z" />
+                <path fill="#de4943" d="M15 4c0.542 0 1 0.458 1 1v4c0 0.542-0.458 1-1 1s-1-0.458-1-1v-4c0-0.542 0.458-1 1-1zM15 2c-1.65 0-3 1.35-3 3v4c0 1.65 1.35 3 3 3s3-1.35 3-3v-4c0-1.65-1.35-3-3-3v0z" />
+                <path fill="#de4943" d="M24 2v7.874c-0.873 0.728-2 1.167-2-1.426v-6.448h-2v6.854h0.003c0.031 1.653 0.379 4.63 3.997 2.21v0.936h2v-10h-2z" />
+                <path fill="#de4943" d="M27 22c-0.551 0-1 0.449-1 1v1h2v-1c0-0.551-0.449-1-1-1z" />
+                <path fill="#de4943" d="M18 23v5.25c0.677 0.677 2 0.75 2-0.25s0-3.688 0-4.688-1-1.313-2-0.313z" />
+                <path fill="#de4943" d="M31.836 18.173c-0.087-1.152-0.501-2.092-1.243-2.82s-1.69-1.125-2.846-1.19c-2.181-0.109-7.17-0.163-11.554-0.163s-9.728 0.054-11.909 0.163c-1.156 0.065-2.104 0.462-2.846 1.19s-1.156 1.668-1.243 2.82c-0.131 2.347-0.196 3.198-0.196 4.372s0.065 2.934 0.196 5.282c0.087 1.152 0.502 2.092 1.243 2.82s1.69 1.125 2.846 1.19c2.181 0.109 7.525 0.163 11.909 0.163s9.373-0.054 11.554-0.163c1.156-0.065 2.104-0.462 2.846-1.19s1.156-1.668 1.243-2.82c0.109-1.956 0.164-3.717 0.164-5.282s-0.055-2.416-0.164-4.372zM6 30h-2v-10h-2v-2h6v2h-2v10zM14 30h-2v-0.936c-3.802 2.134-3.966 0.069-3.997-1.585h-0.003v-5.479h2v5.51c0 1.208 1.127 1.091 2 0.364v-5.874h2v8zM22 27.971c0 2.090-2.141 2.7-4 1.075v0.954h-2v-12h2v3.25c2-2 4-1.25 4 0.75s0 3.787 0 5.971zM30 24v1h-4v2c0 0.551 0.449 1 1 1s1-0.449 1-1v-1h2v1c0 1.654-1.346 3-3 3s-3-1.346-3-3v-4c0-1.654 1.346-3 3-3s3 1.346 3 3v1z" />
+            </svg>
+            <svg id="icon-vimeo2" width="32" height="32" viewBox="0 0 32 32" x="0" y="210">
+                <path fill="#04ae7f" d="M26.666 0h-21.332c-2.934 0-5.334 2.401-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.332c2.934 0 5.334-2.398 5.334-5.334v-21.332c0-2.934-2.4-5.334-5.334-5.334zM25.402 13.691c-1.061 5.978-6.991 11.041-8.774 12.197s-3.411-0.463-4.001-1.687c-0.675-1.397-2.7-8.967-3.23-9.594-0.53-0.626-2.121 0.627-2.121 0.627l-0.772-1.012c0 0 3.231-3.858 5.689-4.34 2.606-0.511 2.603 4.002 3.229 6.508 0.606 2.424 1.014 3.811 1.543 3.811 0.53 0 1.543-1.352 2.652-3.424 1.111-2.074-0.048-3.907-2.218-2.603 0.868-5.208 9.064-6.46 8.004-0.483z" />
+            </svg>
+            <svg id="icon-lanyrd" width="32" height="32" viewBox="0 0 32 32" x="0" y="252">
+                <path fill="#ef6db3" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM26.365 25.24c-3.542 1.167-10.458 3.438-11.75 3.76s-2.625 0.865-3.375-1.302c0 0-6.375-18.833-6.958-20.5s-1.151-2.671 0.167-3.198c3.043-1.215 6.208-2.766 6.792-1 0.522 1.58 6.083 19 6.083 19s6.542-2.302 8.333-2.802 2.406-0.833 2.875 1.625 1.375 3.25-2.167 4.417z" />
+            </svg>
+            <svg id="icon-flickr3" width="32" height="32" viewBox="0 0 32 32" x="0" y="294">
+                <path fill="#e31884" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-21.332c0-2.933-2.399-5.334-5.332-5.334zM9 21c-2.761 0-5-2.239-5-5s2.239-5 5-5 5 2.239 5 5c0 2.761-2.239 5-5 5zM23 21c-2.761 0-5-2.239-5-5s2.239-5 5-5 5 2.239 5 5c0 2.761-2.239 5-5 5z" />
+            </svg>
+            <svg id="icon-picassa2" width="32" height="32" viewBox="0 0 32 32" x="0" y="336">
+                <path fill="#fa8526" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-21.332c0-2.933-2.399-5.334-5.332-5.334zM16 2c2.147 0 4.181 0.484 6 1.348v8.973l-9.799-9.799c1.208-0.34 2.482-0.522 3.799-0.522zM2 16c0-5.497 3.168-10.252 7.777-12.543l3.543 3.543-11.155 11.155c-0.109-0.703-0.165-1.422-0.165-2.155zM8 27.489c-2.334-1.628-4.149-3.948-5.146-6.664l5.146-5.146v11.81zM16 30c-2.147 0-4.181-0.484-6-1.348v-4.652h17.489c-2.53 3.627-6.732 6-11.489 6zM28.652 22h-4.652v-17.489c3.627 2.53 6 6.732 6 11.489 0 2.147-0.484 4.181-1.348 6z" />
+            </svg>
+            <svg id="icon-dribbble2" width="32" height="32" viewBox="0 0 32 32" x="0" y="378">
+                <path fill="#ef6797" d="M16.836 16.123c-6.167 1.994-9.449 7.427-9.72 7.891-1.91-2.12-3.074-4.924-3.074-7.996 0-0.122 0.002-0.244 0.006-0.366 0.523 0.011 6.318 0.085 12.29-1.702 0.342 0.67 0.669 1.35 0.97 2.029-0.158 0.045-0.315 0.092-0.471 0.142zM23.927 7.073c-0.215 0.306-2.015 2.726-6.267 4.32-1.96-3.6-4.131-6.558-4.462-7 0.899-0.217 1.837-0.332 2.802-0.332 3.038 0 5.815 1.14 7.927 3.013zM15.35 12.112c-5.594 1.485-10.52 1.464-11.055 1.457 0.775-3.711 3.276-6.795 6.622-8.373 0.315 0.432 2.452 3.394 4.434 6.916zM18.041 17.984c0.026-0.009 0.052-0.018 0.079-0.027 1.679 4.363 2.373 8.019 2.551 9.068-1.436 0.611-3.015 0.95-4.671 0.95-2.764 0-5.313-0.943-7.34-2.524 0.215-0.44 2.636-5.117 9.381-7.468zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM16 30c-7.72 0-14-6.28-14-14s6.28-14 14-14c7.72 0 14 6.28 14 14s-6.28 14-14 14zM20.358 17.405c3.747-0.6 7.040 0.382 7.449 0.511-0.533 3.327-2.443 6.201-5.126 8.015-0.123-0.727-0.759-4.228-2.323-8.526zM18.579 13.171c4.421-1.805 6.425-4.374 6.677-4.715 1.665 2.035 2.674 4.627 2.7 7.449-0.394-0.083-4.348-0.885-8.327-0.385-0.084-0.205-0.17-0.41-0.258-0.617-0.248-0.583-0.515-1.162-0.793-1.732z" />
+            </svg>
+            <svg id="icon-forrst2" width="32" height="32" viewBox="0 0 32 32" x="0" y="420">
+                <path fill="#d61e1e" d="M14 18h4v8h-4zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM18 26v4h-4v-4h-8l10-22 10 22h-8z" />
+            </svg>
+            <svg id="icon-deviantart2" width="32" height="32" viewBox="0 0 32 32" x="0" y="462">
+                <path fill="#0b6e5c" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM30 24h-12v-7.375c0 0 2.198 0.031 3.375 1.25v2.125h4.825c-0.369-1.184-1.158-2.347-2.285-3.333-1.937-1.695-5.1-2.542-7.915-2.542v9.875h-14v-3c0-1.356 0.458-2.704 1.324-3.899 0.769-1.060 1.84-1.995 3.183-2.779 1.576-0.919 3.455-1.585 5.493-1.962v-4.361h4v4c4.594 0 6.925 0.825 9.493 2.323 1.344 0.784 2.415 1.719 3.183 2.779 0.866 1.194 1.324 2.543 1.324 3.899v3zM6.237 20c0.369-1.184 1.158-2.347 2.285-3.333 0.986-0.862 2.176-1.537 3.478-1.992v5.326h-5.763z" />
+            </svg>
+            <svg id="icon-steam2" width="32" height="32" viewBox="0 0 32 32" x="0" y="504">
+                <path fill="#c45d3e" d="M9.498 26.125c0.848 0 1.681-0.41 2.187-1.169 0.804-1.206 0.478-2.836-0.728-3.64l-2.065-1.377c0.353-0.096 0.725-0.149 1.109-0.149 2.325 0 4.209 1.884 4.209 4.209s-1.885 4.209-4.209 4.209c-2.263 0-4.108-1.785-4.205-4.024l2.249 1.499c0.448 0.298 0.953 0.441 1.454 0.441zM26.667 0c2.934 0 5.333 2.4 5.333 5.334v21.332c0 2.936-2.399 5.334-5.333 5.334h-21.333c-2.934 0-5.334-2.398-5.334-5.334v-6.345l3.792 2.528c-0.364 1.974 0.217 4.088 1.744 5.615 2.466 2.466 6.463 2.466 8.929 0 1.515-1.515 2.098-3.607 1.752-5.568l7.783-6.968c1.445-0.207 2.838-0.867 3.95-1.979 2.734-2.734 2.734-7.166 0-9.899s-7.166-2.734-9.9 0c-1.112 1.112-1.771 2.505-1.979 3.95v0l-6.397 9.695c-0.995 0.051-1.981 0.336-2.866 0.856l-6.808-4.539v-8.678c0-2.934 2.4-5.334 5.334-5.334h21.333zM28 9c0-2.761-2.239-5-5-5s-5 2.239-5 5 2.239 5 5 5 5-2.239 5-5zM20 9c0-1.657 1.343-3 3-3s3 1.343 3 3-1.343 3-3 3-3-1.343-3-3z" />
+            </svg>
+            <svg id="icon-github" width="32" height="32" viewBox="0 0 32 32" x="0" y="546">
+                <path fill="#484848" d="M16 0.395c-8.836 0-16 7.163-16 16 0 7.069 4.585 13.067 10.942 15.182 0.8 0.148 1.094-0.347 1.094-0.77 0-0.381-0.015-1.642-0.022-2.979-4.452 0.968-5.391-1.888-5.391-1.888-0.728-1.849-1.776-2.341-1.776-2.341-1.452-0.993 0.11-0.973 0.11-0.973 1.606 0.113 2.452 1.649 2.452 1.649 1.427 2.446 3.743 1.739 4.656 1.33 0.143-1.034 0.558-1.74 1.016-2.14-3.554-0.404-7.29-1.777-7.29-7.907 0-1.747 0.625-3.174 1.649-4.295-0.166-0.403-0.714-2.030 0.155-4.234 0 0 1.344-0.43 4.401 1.64 1.276-0.355 2.645-0.532 4.005-0.539 1.359 0.006 2.729 0.184 4.008 0.539 3.054-2.070 4.395-1.64 4.395-1.64 0.871 2.204 0.323 3.831 0.157 4.234 1.026 1.12 1.647 2.548 1.647 4.295 0 6.145-3.743 7.498-7.306 7.895 0.574 0.497 1.085 1.47 1.085 2.963 0 2.141-0.019 3.864-0.019 4.391 0 0.426 0.288 0.925 1.099 0.768 6.354-2.118 10.933-8.113 10.933-15.18 0-8.837-7.164-16-16-16z" />
+            </svg>
+            <svg id="icon-wordpress2" width="32" height="32" viewBox="0 0 32 32" x="0" y="588">
+                <path fill="#4a4a4a" d="M16.211 17.022l3.688 9.842c0.024 0.058 0.054 0.111 0.086 0.161-1.247 0.428-2.588 0.663-3.985 0.663-1.178 0-2.315-0.168-3.39-0.476l3.601-10.189zM28 16c0 4.312-2.399 8.076-5.967 10.102l3.665-10.321c0.685-1.667 0.913-3.001 0.913-4.186 0-0.43-0.029-0.83-0.081-1.202 0.937 1.664 1.47 3.575 1.47 5.607zM4 16c0-1.694 0.373-3.303 1.039-4.756l5.724 15.274c-4.003-1.894-6.763-5.892-6.763-10.518zM5.974 9.58c2.146-3.173 5.834-5.268 10.026-5.268 3.124 0 5.968 1.163 8.103 3.069-0.052-0.003-0.102-0.009-0.155-0.009-1.179 0-2.015 1-2.015 2.074 0 0.963 0.571 1.778 1.179 2.741 0.456 0.778 0.989 1.778 0.989 3.223 0 1-0.395 2.161-0.913 3.778l-1.197 3.895-4.337-12.563c0.722-0.037 1.373-0.111 1.373-0.111 0.647-0.075 0.571-1-0.076-0.963 0 0-1.944 0.149-3.198 0.149-1.179 0-3.16-0.149-3.16-0.149-0.647-0.037-0.723 0.926-0.076 0.963 0 0 0.612 0.074 1.259 0.111l1.869 4.989-2.626 7.671-4.37-12.66c0.723-0.037 1.373-0.111 1.373-0.111 0.646-0.075 0.57-1-0.077-0.963 0 0-1.943 0.149-3.198 0.149-0.225 0-0.491-0.005-0.772-0.014zM26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM30 16c0 7.732-6.268 14-14 14s-14-6.268-14-14 6.268-14 14-14 14 6.268 14 14z" />
+            </svg>
+            <svg id="icon-blogger2" width="32" height="32" viewBox="0 0 32 32" x="0" y="630">
+                <path fill="#f17c1d" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.388 5.334 5.333 5.334h21.334c2.945 0 5.333-2.388 5.333-5.334l0-21.332c0-2.946-2.388-5.334-5.333-5.334zM28 20.253c0 4.281-3.484 7.747-7.795 7.747h-8.401c-4.308 0-7.804-3.466-7.804-7.747v-8.503c0.003-4.281 3.496-7.75 7.804-7.75h3.933c4.311 0 7.762 3.219 7.762 7.5 0.056 0.803 0.779 1.5 1.603 1.5h1.345c0.862 0 1.552 0.904 1.552 1.76v5.493zM22 20c0 1.1-0.9 2-2 2h-8c-1.1 0-2-0.9-2-2v0c0-1.1 0.9-2 2-2h8c1.1 0 2 0.9 2 2v0zM18 12c0 1.1-0.9 2-2 2h-4c-1.1 0-2-0.9-2-2v0c0-1.1 0.9-2 2-2h4c1.1 0 2 0.9 2 2v0z" />
+            </svg>
+            <svg id="icon-tumblr2" width="32" height="32" viewBox="0 0 32 32" x="0" y="672">
+                <path fill="#3c576d" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.334c2.933 0 5.332-2.398 5.332-5.334v-21.332c-0-2.933-2.399-5.334-5.332-5.334zM22.866 25.771c-0.942 0.443-1.798 0.756-2.563 0.936-0.765 0.178-1.593 0.267-2.481 0.267-1.010 0-1.605-0.127-2.381-0.381-0.775-0.256-1.438-0.621-1.984-1.090-0.549-0.473-0.928-0.975-1.14-1.506s-0.317-1.303-0.317-2.313v-7.744h-3v-3.127c0.867-0.281 1.873-0.685 2.49-1.211 0.62-0.527 1.116-1.158 1.49-1.896 0.375-0.736 0.633-1.676 0.774-2.815h3.141v5.108h5.105v3.941h-5.106v5.662c0 1.281-0.017 2.020 0.119 2.383 0.135 0.361 0.473 0.736 0.841 0.953 0.489 0.293 1.047 0.439 1.676 0.439 1.118 0 2.231-0.363 3.336-1.090v3.482z" />
+            </svg>
+            <svg id="icon-soundcloud2" width="32" height="32" viewBox="0 0 32 32" x="0" y="714">
+                <path fill="#dbbaa8" d="M26.668 0h-21.334c-2.934 0-5.334 2.4-5.334 5.334v14.666l0.5-2h1l0.5 2-0.5 2h-1l-0.5-2v6.666c0 2.936 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.399 5.332-5.334v-6.631c0 2.19-1.846 3.965-4.125 3.965-0.074 0-11.424-0.006-11.434-0.007-0.246-0.024-0.441-0.226-0.441-0.469v-12.585c0-0.231 0.082-0.35 0.393-0.466 0.801-0.298 1.704-0.474 2.635-0.474 3.803 0 6.922 2.803 7.25 6.377 0.491-0.198 1.031-0.308 1.596-0.308 2.278 0 4.126 1.776 4.126 3.966v-14.701c0-2.934-2.399-5.334-5.332-5.334zM5.5 24h-1l-0.5-4 0.5-4h1l0.5 4-0.5 4zM9.5 24h-1l-0.5-5.086 0.5-4.914h1l0.5 5-0.5 5zM13.5 24h-1l-0.5-7.007 0.5-6.993h1l0.5 6.993-0.5 7.007z" />
+            </svg>
+            <svg id="icon-linkedin" width="32" height="32" viewBox="0 0 32 32" x="0" y="756">
+                <path fill="#3c91b3" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM12 26h-4v-14h4v14zM10 10c-1.105 0-2-0.895-2-2s0.895-2 2-2 2 0.895 2 2-0.895 2-2 2zM26 26h-4v-8c0-1.105-0.895-2-2-2s-2 0.895-2 2v8h-4v-14h4v2.483c0.825-1.133 2.086-2.483 3.5-2.483 2.485 0 4.5 2.239 4.5 5v9z" />
+            </svg>
+            <svg id="icon-lastfm2" width="32" height="32" viewBox="0 0 32 32" x="0" y="798">
+                <path fill="#b8bb8c" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM23.329 23.778c-5.549 0-7.473-2.502-8.499-5.613l-1.027-3.207c-0.77-2.341-1.668-4.169-4.49-4.169-1.957 0-3.945 1.411-3.945 5.356 0 3.079 1.572 5.003 3.785 5.003 2.501 0 4.169-1.86 4.169-1.86l1.027 2.79c0 0-1.732 1.7-5.356 1.7-4.49 0-6.992-2.63-6.992-7.505 0-5.068 2.502-8.050 7.216-8.050 4.266 0 6.415 1.54 7.762 5.709l1.058 3.207c0.77 2.342 2.117 4.041 5.356 4.041 2.181 0 3.335-0.481 3.335-1.668 0-0.93-0.545-1.604-2.181-1.989l-2.181-0.513c-2.662-0.641-3.72-2.021-3.72-4.202 0-3.496 2.823-4.587 5.709-4.587 3.272 0 5.26 1.187 5.517 4.073l-3.207 0.385c-0.128-1.379-0.962-1.957-2.502-1.957-1.411 0-2.277 0.642-2.277 1.732 0 0.962 0.417 1.54 1.828 1.86l2.053 0.449c2.758 0.642 4.234 1.989 4.234 4.587-0 3.207-2.694 4.426-6.672 4.426z" />
+            </svg>
+            <svg id="icon-delicious" width="32" height="32" viewBox="0 0 32 32" x="0" y="840">
+                <path fill="#b1110f" d="M26.668 0h-21.334c-2.934 0-5.334 2.401-5.334 5.334v21.332c0 2.935 2.4 5.334 5.334 5.334h21.334c2.934 0 5.332-2.398 5.332-5.334v-21.332c0-2.934-2.398-5.334-5.332-5.334zM30 26.666c0 0.885-0.348 1.721-0.98 2.354s-1.467 0.98-2.352 0.98h-10.668v-14h-14v-10.666c0-0.885 0.348-1.72 0.981-2.353s1.468-0.981 2.353-0.981h10.666v14h14v10.666z" />
+            </svg>
+            <svg id="icon-stumbleupon" width="32" height="32" viewBox="0 0 32 32" x="0" y="882">
+                <path fill="#60bd0d" d="M26.625 0h-21.25c-2.956 0-5.375 2.419-5.375 5.375v21.25c0 2.956 2.419 5.375 5.375 5.375h21.25c2.956 0 5.375-2.419 5.375-5.375v-21.25c0-2.956-2.419-5.375-5.375-5.375zM16 10c-1.103 0-2 0.897-2 2v8c0 3.308-2.692 6-6 6s-6-2.692-6-6v-4h4v4c0 1.103 0.897 2 2 2s2-0.897 2-2v-8c0-3.308 2.692-6 6-6s6 2.692 6 5.563v1.938l-2.563 0.75-1.438-0.75v-1.938c0-0.665-0.897-1.563-2-1.563zM30 20c0 3.308-2.692 6-6 6s-6-2.692-6-6.438v-3.875l1.438 0.75 2.563-0.75v3.875c0 1.54 0.897 2.438 2 2.438s2-0.897 2-2v-4h4v4z" />
+            </svg>
+            <svg id="icon-pinterest2" width="32" height="32" viewBox="0 0 32 32" x="0" y="924">
+                <path fill="#2f6e95" d="M26.667 0h-21.334c-2.945 0-5.333 2.388-5.333 5.334v21.332c0 2.946 2.387 5.334 5.333 5.334h21.334c2.945 0 5.333-2.388 5.333-5.334v-21.332c0-2.946-2.387-5.334-5.333-5.334zM17.915 25.126c-1.621 0-3.145-0.842-3.667-1.837 0 0-0.802 3.055-0.997 3.803-0.361 1.39-1.336 3.132-1.989 4.195l-1.093-0.387c-0.14-1.266-0.266-3.208 0.055-4.59 0.291-1.249 1.876-7.953 1.876-7.953s-0.479-0.958-0.479-2.375c0-2.225 1.29-3.886 2.895-3.886 1.365 0 2.025 1.025 2.025 2.254 0 1.373-0.874 3.425-1.325 5.327-0.377 1.593 0.799 2.892 2.369 2.892 2.844 0 5.030-2.999 5.030-7.327 0-3.831-2.753-6.509-6.683-6.509-4.552 0-7.225 3.415-7.225 6.943 0 1.375 0.53 2.85 1.191 3.651 0.131 0.158 0.15 0.297 0.111 0.459-0.121 0.506-0.391 1.593-0.444 1.815-0.070 0.293-0.232 0.355-0.535 0.214-1.998-0.93-3.248-3.852-3.248-6.198 0-5.047 3.667-9.682 10.572-9.682 5.55 0 9.864 3.955 9.864 9.241 0 5.514-3.477 9.952-8.302 9.952z" />
+            </svg>
+            <svg id="icon-xing" width="32" height="32" viewBox="0 0 32 32" x="0" y="966">
+                <path fill="#914e93" d="M26.667 0h-21.333c-2.934 0-5.334 2.4-5.334 5.334v21.332c0 2.936 2.4 5.334 5.334 5.334h21.333c2.934 0 5.333-2.398 5.333-5.334v-21.332c0-2.934-2.399-5.334-5.333-5.334zM9.743 20.837h-3.477c-0.209 0-0.366-0.094-0.454-0.237-0.092-0.149-0.096-0.341 0-0.533l3.694-6.52c0.004-0.007 0.004-0.012 0-0.020l-2.351-4.070c-0.097-0.194-0.111-0.385-0.020-0.533 0.089-0.144 0.265-0.217 0.474-0.217h3.477c0.533 0 0.795 0.344 0.968 0.652 0 0 2.376 4.146 2.391 4.168-0.141 0.248-3.754 6.638-3.754 6.638-0.178 0.321-0.428 0.672-0.948 0.672zM26.185 4.704l-7.698 13.612c-0.005 0.007-0.005 0.016 0 0.023l4.902 8.956c0.097 0.193 0.1 0.389 0.007 0.537-0.088 0.143-0.252 0.222-0.461 0.222h-3.473c-0.533 0-0.799-0.354-0.971-0.662 0 0-4.928-9.040-4.941-9.064 0.246-0.436 7.737-13.72 7.737-13.72 0.186-0.335 0.412-0.662 0.931-0.662h3.513c0.209 0 0.374 0.079 0.462 0.222 0.092 0.149 0.089 0.344-0.007 0.537z" />
+            </svg>
+            <svg id="icon-flattr" width="32" height="32" viewBox="0 0 32 32" x="0" y="1008">
+                <path fill="#b9c5bf" d="M11.486 0c-7.605 0-11.486 4.38-11.486 12.558v0 17.157l7.45-7.457v-8.716c0-3.388 0.898-5.544 3.91-6.028v0c1.052-0.206 3.242-0.134 4.635-0.134v0 5.175c0 0.047 0.006 0.131 0.019 0.175v0c0.058 0.21 0.248 0.363 0.472 0.363v0c0.127 0 0.246-0.066 0.368-0.187v0l12.91-12.905-18.277-0.002zM24.55 9.742v8.715c0 3.388-0.898 5.544-3.91 6.028v0c-1.052 0.206-3.242 0.134-4.635 0.134v0-5.175c0-0.047-0.006-0.131-0.019-0.174v0c-0.058-0.21-0.248-0.363-0.472-0.363v0c-0.127-0-0.246 0.066-0.368 0.186v0l-12.91 12.905 18.277 0.002c7.605 0 11.486-4.38 11.486-12.558v0-17.157l-7.45 7.457z" />
+            </svg>
+        </svg>
+    </div>
 </template>
 
 <style lang="scss">
-    ul {
-        // counter-reset: svg;
+ul {
+    // counter-reset: svg;
+}
+
+.svg {
+    width: 0.32rem;
+    height: 0.32rem;
+    position: relative;
+    &:after {
+        // counter-increment: svg;
+        // content: counter(svg);
+        content: '';
+        position: absolute;
+        left: 0;
+        top: 0;
+        width: 100%;
+        height: 100%;
+        background: url(http://www.w3cplus.com/sites/default/files/blogs/2015/1506/sprite.svg) no-repeat;
+        // background-position: 0 counter(svg)*0.42rem;
     }
-    .svg {
-        width: 0.32rem;
-        height: 0.32rem;
-        position: relative;
+}
 
-        &:after {
-            // counter-increment: svg;
-            // content: counter(svg);
-            content: '';
-            position: absolute;
-            left: 0;
-            top: 0;
-            width: 100%;
-            height: 100%;
-            background: url(http://www.w3cplus.com/sites/default/files/blogs/2015/1506/sprite.svg) no-repeat;
-            // background-position: 0 counter(svg)*0.42rem;
-        }
-    }
-
-    // @for $i from 1 through 10 {
-    //     .svg:nth-of-type(#{$i})
-    // }
-
+// @for $i from 1 through 10 {
+//     .svg:nth-of-type(#{$i})
+// }
 </style>
